PART4: Code các chức năng cho OrderDetail

+++ SHOPAPP-BACKEND/shopapp/src/main/java/com/project/shopapp/dtos/OrderDetailDTO.java
@@ -20,7 +20,7 @@ public class OrderDetailDTO {
     private long productId;
 
     @Min(value = 0, message = "Price must be greater than or equal to 0")
-    private long price;
+    private Float price;
 
     @JsonProperty("number_of_products")
     @Min(value = 1, message = "Number of products must be greater than or equal to 1")
